requirements.txt.txt
Pillow==9.2.0
langchain==0.1.17
langchain-experimental==0.0.57
langchain-google-genai==1.0.3
gradio==3.41.2
gradio_client==0.5.0